Create Define PI Ratio Program
This code takes the output from the 1x128 DNN to calculate the number of visits predicted to be 'high risk' vs. the total number of visits for a given individual.  It then ranks and outputs a list of health plan members in descending order.  Total calls for the batch run are also displayed.  Note, this program will perform the predictions with the embedded DNN model but you must have the model (i.e. 'spirit_dnn.sav') which is specified by the "spirit_dnn1x128" variable here.